[CMH] 일지 화면 구현 작업 중
