+++ teamproject/src/main/resources/mapper/WorkoutLogConfigMapper.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+        
+<mapper namespace="aihometraining.team.workoutLog.mapper.WorkoutLogConfigMapper">
+
+	<resultMap type="WorkoutLogCategory" id="WorkoutLogCategoryResultMap">
+		<result column="workoutGoalPlanCategoryCode" 		property="workoutGoalPlanCategoryCode"/>
+		<result column="workoutGoalPlanCategoryContent" 	property="workoutGoalPlanCategoryContent"/>
+	</resultMap>
+	
+	<select id="getWorkoutLogCategoryList" resultType="WorkoutLogCategory">
+		/* 운동 카테고리 조회 */
+		SELECT
+			workoutGoalPlanCategoryContent
+		FROM
+			workoutgoalplancategory;
+	</select>
+</mapper>
