+++ teamproject/src/main/resources/templates/userfragments/header.html
@@ -43,9 +43,10 @@
                                             </li>
                                             <li><a href="#">#일지</a>
                                                 <ul class="submenu">
-                                                    <li><a th:href="@{/workoutLog/workoutLogUser/workoutGoalPlanList}">#목표 및 계획(민희)</a></li>
+                                                    <li><a th:href="@{/workoutLog/workoutLogUser/workoutGoalList}">#운동 목표(민희)</a></li>
+                                                    <li><a th:href="@{/workoutLog/workoutLogUser/workoutGoalPlanList}">#운동 계획(민희)</a></li>
                                                     <li><a th:href="@{/workoutLog/workoutLogUser/workoutLogMain}">#하루 일지(민희)</a></li>
-                                                    <li><a th:href="@{/workoutLog/workoutLogUser/workoutLogList}">#인기 일지(민희)</a></li>
+                                                    <li><a href="#">#인기 일지(민희)</a></li>
                                                 </ul>
                                             </li>
                                             <li><a th:href="@{/wishList}">#위시리스트</a></li>
