+++ teamproject/src/main/resources/templates/workoutLog/workoutLogUser/workoutGoalDelete.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org"
 	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-	  layout:decorate="~{layout/default}">
+	  layout:decorate="~{layout/userdefault}">
 	<!-- 사용자 정의 title -->
 	<th:block layout:fragment="customTitle">
 		<title th:text="${title}"></title>
@@ -11,21 +11,5 @@
 		일지 사용자 관리 화면입니다.
 	
 	</th:block>
-		<th:block layout:fragment="customBodyScript">	
-		<!-- plugins -->
-		<script th:src="@{/asset/js/plugins/moment.min.js}"></script>
-		<script th:src="@{/asset/js/plugins/jquery.datatables.min.js}"></script>
-		<script th:src="@{/asset/js/plugins/datatables.bootstrap.min.js}"></script>
-		<script th:src="@{/asset/js/plugins/jquery.nicescroll.js}"></script>
-		
-		
-		<!-- custom -->
-		<script th:src="@{/asset/js/main.js}"></script>
-		<script type="text/javascript">
-		  $(document).ready(function(){
-		    //$('#datatables-example').DataTable();
-		  });
-		</script>
-<!-- end: Javascript -->
-	</th:block>
+	
 </html>
